requirements.txt
python-telegram-bot==20.7
gradio-client==0.10.1
python-dotenv==1.0.0 